chore

+++ src/roq/okx/market_data.hpp
@@ -48,7 +48,7 @@ class MarketData final : public web::socket::Client::Handler, public json::Parse
     virtual void operator()(SymbolsUpdate &) = 0;
   };
 
-  MarketData(Handler &, io::Context &, uint32_t stream_id, Security &, Shared &, size_t index);
+  MarketData(Handler &, io::Context &, uint16_t stream_id, Security &, Shared &, size_t index);
 
   MarketData(MarketData &&) = delete;
   MarketData(MarketData const &) = delete;
